Makefile

+++ admin/scripts/get-scripts.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 
-wget -O marked.js https://github.com/markedjs/marked/raw/master/marked.min.js
+[ -f marked.js ] && exit 0
 
+wget -O marked.js https://github.com/markedjs/marked/raw/master/marked.min.js
